Python-Choosing-Giveaway-Winner-CSV
I made this python program to select 10 winners from data I collected from a contest I conducted on my gaming YouTube channel.

Data was collected using various parameters and the program increases the ODDS of participants based on certain conditions example: if the user has subscribed to Amogh's YouTube Channel or not, If they have, it increases the odds by 10%..

once program is executed, It generates a winners.csv file.